+++ RELEASE_NOTES.md
@@ -1,1 +1,3 @@
 # ChefDK 1.1 Release notes
+
+* Adds 'chef generate helpers HELPERS_NAME' to generate a helpers file in libraries